4.1/devops/lab3/lab3/task3/requirements.txt
flask>=2.0
opentelemetry-distro
opentelemetry-instrumentation-flask
opentelemetry-exporter-otlp-proto-grpc>=1.26.0
opentelemetry-exporter-otlp>=1.26.0
prometheus-client